# some microoptimizations for the resolver