--- a/src/Resolve/versionweights.jl
+++ b/src/Resolve/versionweights.jl
@@ -25,19 +25,12 @@ Base.:(+)(a::VersionWeight, b::VersionWeight) =
 Base.:(-)(a::VersionWeight) =
     VersionWeight(-a.major, -a.minor, -a.patch)
 
-function Base.cmp(a::VersionWeight, b::VersionWeight)
-    c = cmp(a.major, b.major); c != 0 && return c
-    c = cmp(a.minor, b.minor); c != 0 && return c
-    return cmp(a.patch, b.patch)
+function Base.isless(a::VersionWeight, b::VersionWeight)
+    (a.major, a.minor, a.patch) < (b.major, b.minor, b.patch)
 end
-Base.isless(a::VersionWeight, b::VersionWeight) = cmp(a,b) < 0
-Base.:(==)(a::VersionWeight, b::VersionWeight) = cmp(a,b) == 0
 
 Base.abs(a::VersionWeight) =
     VersionWeight(abs(a.major), abs(a.minor), abs(a.patch))
-
-Base.copy(a::VersionWeight) =
-    VersionWeight(a.major, a.minor, a.patch)
 
 # This isn't nice, but it's for debugging only anyway
 function Base.show(io::IO, a::VersionWeight)