--- a/src/Types.jl
+++ b/src/Types.jl
@@ -19,7 +19,7 @@ import ..PlatformEngines: probe_platform_engines!, download, download_verify_unp
 import Base: SHA1
 using SHA
 
-export UUID, pkgID, SHA1, VersionRange, VersionSpec,
+export UUID, SHA1, VersionRange, VersionSpec,
     PackageSpec, EnvCache, Context, PackageInfo, ProjectInfo, GitRepo, Context!, range_compressed_versionspec, err_rep,
     PkgError, pkgerror, has_name, has_uuid, is_stdlib, stdlibs, write_env, write_env_usage, parse_toml, find_registered!,
     project_resolve!, project_deps_resolve!, manifest_resolve!, registry_resolve!, stdlib_resolve!, handle_repos_develop!, handle_repos_add!, ensure_resolved,
@@ -35,13 +35,6 @@ export UUID, pkgID, SHA1, VersionRange, VersionSpec,
 include("versions.jl")
 
 const URL_regex = r"((file|git|ssh|http(s)?)|(git@[\w\-\.]+))(:(//)?)([\w\.@\:/\-~]+)(\.git)?(/)?"x
-
-## user-friendly representation of package IDs ##
-function pkgID(p::UUID, uuid_to_name::Dict{UUID,String})
-    name = get(uuid_to_name, p, "(unknown)")
-    uuid_short = string(p)[1:8]
-    return "$name [$uuid_short]"
-end
 
 #################
 # Pkg Error #