These are results for `Pkg.update` in an environment where I added a bunch of big packages (Plots, Makie, DifferentialEquations, ApproxFun etc).


Master:

```
ENV["JULIA_PKGRESOLVE_ACCURACY"] = 1
  966.591 ms (6843000 allocations: 1.01 GiB)

ENV["JULIA_PKGRESOLVE_ACCURACY"] = 2
  2.124 s (11577823 allocations: 3.01 GiB)
```


PR:

```
ENV["JULIA_PKGRESOLVE_ACCURACY"] = 1
  602.627 ms (5432062 allocations: 301.72 MiB)

ENV["JULIA_PKGRESOLVE_ACCURACY"] = 2
   1.308 s (7230777 allocations: 368.51 MiB)
```

The performance improvements here only matter when the greedy solver fails.

The reason for this is that https://github.com/JuliaLang/Pkg.jl/issues/1949 suggests that we might want to bump the resolver accuracy but it would be a bit sad if we lost performance from doing that. The optimized resolver is now pretty close for accuracy 2, as the previous one with accuracy 1.
 
cc @carlobaldassi