@carlobaldassi 2020-08-12
This all looks good to me, thanks.
As for increasing the default accuracy, it might be that with the growing ecosystem and larger and more complicated dependency graphs it's going to be necessary. I should think if there is any way we can come up with a better heuristic to set it, but at the moment nothing particularly clever comes to mind.